Move test checker to src/
