+++ pkgs/intl4x/lib/src/collator/collator_impl.dart
@@ -4,7 +4,7 @@
 
 import '../../intl4x.dart';
 
-import '../../intl4x_test_checker.dart';
+import '../intl4x_test_checker.dart';
 import 'collator_options.dart';
 
 abstract class Collator {
