+++ pkgs/intl4x/lib/src/list_format/list_formatter.dart
@@ -4,7 +4,7 @@
 
 import '../../intl4x.dart';
 
-import '../../intl4x_test_checker.dart';
+import '../intl4x_test_checker.dart';
 
 abstract class ListFormatter {
   final Intl intl;
