+++ pkgs/intl4x/lib/src/number_format/number_formatter.dart
@@ -4,7 +4,7 @@
 
 import '../../intl4x.dart';
 
-import '../../intl4x_test_checker.dart';
+import '../intl4x_test_checker.dart';
 
 abstract class NumberFormatter {
   final Intl intl;
